- removed url from base and put in secure.py - removed duplicate and unused methods - put pin service back in its own folder structure - put masquerade service back in its own folder structure - moved "masquerade_confirm" as part of "masquerade_as" function
