+++ selenium_tests/masquerade/canvas_base_page_object.py
@@ -1,0 +1,7 @@
+# from selenium_tests.base_page_object import BasePageObject
+#
+#
+# class CanvasBasePageObject(BasePageObject):
+#     pass
+#     # def __init__(self, *args, **kwargs):
+#     #     super(CanvasBasePageObject, self).__init__(*args, **kwargs)
